requirements-gpu.txt
setuptools==41.0.1
wheel==0.33.4
pytest==3.9.2
tensorflow-gpu==1.13.1
chardet==3.0.4
nltk==3.2.1
beautifulsoup4==4.6.0
